+++ Tests/AnomaliesTestCase.swift
@@ -10,8 +10,6 @@ import XCTest
 import DTTableViewManager
 import DTModelStorage
 
-#if swift(>=4.1)
-
 enum TestAnomaly : Equatable, CustomDebugStringConvertible {
     case itemEventCalledWithCellType(ObjectIdentifier)
     case weirdIndexPathAction(IndexPath)
@@ -63,5 +61,3 @@ class AnomaliesTestCase: XCTestCase {
         waitForExpectations(timeout: 0.1)
     }
 }
-
-#endif