+++ Tests/DatasourceTestCase.swift
@@ -302,7 +302,6 @@ class DatasourceTestCase: XCTestCase {
         waitForExpectations(timeout: 0.5, handler: nil)
     }
     
-#if swift(>=4.1)
     func testNilModelInStorageLeadsToNilModelAnomaly() {
         let exp = expectation(description: "Nil model in storage")
         let model: Int?? = nil
@@ -478,5 +477,4 @@ class DatasourceTestCase: XCTestCase {
         
         XCTAssertEqual(anomaly.debugDescription, "⚠️[DTTableViewManager] Attempted to register xib NibView, but view found in a xib was of type NibView, while expected type is ReactingHeaderFooterView. This can prevent headers/footers from being updated with models and react to events.")
     }
-#endif
 }