+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
+## [6.5.0](https://github.com/DenTelezhkin/DTTableViewManager/releases/tag/6.5.0)
+
 ### Added
 
 * Convenience constructor for `DTTableViewManager` object: `init(storage:)` that allows to create it's instance without initializing `MemoryStorage`.